pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "COVE"
version = "0.0.1"
authors = [
  { name="Ryan DeWolfe", email="ryandewolfe33@gmail.com" },
]
description = "Co-Occurance Vertex Embedding"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E*"]

requires-python = ">=3.10"
dependencies = [
    "numpy >= 2.0",
    "scipy >= 1.16",
    "numba >= 0.60.0",
]

[project.optional-dependencies]
dr = [
    "umap-learn >= 0.5",
    "scikit-learn",
]

[tool.setuptools]
packages = ["cove"]
